+++ .vscode/launch.json
@@ -10,7 +10,8 @@
             "request": "launch",
             "module": "uvicorn",
             "args": [
-                "app.main:app"
+                "app.main:app",
+                "--reload"
             ],
             "jinja": true,
             "justMyCode": true
