+++ app/crud.py
@@ -55,7 +55,7 @@ def get_group_by_name(db:Session,groupname:str):
         return None
         '''
 
-def create_user(db:Session,user:schemas.user.UserCreate):
+def create_user(db:Session,user:schemas.UserCreate):
     hashed_password = dep.get_password_hash(user.password)
     db_user = models.User(username=user.username, is_family=False,is_active=False,password_expired=False,hashed_password=hashed_password)
     db.add(db_user)
@@ -63,7 +63,7 @@ def create_user(db:Session,user:schemas.user.UserCreate):
     db.refresh(db_user)
     return db_user
 
-def create_user_by_admin(db:Session,user:schemas.user.UserCreateByAdmin):
+def create_user_by_admin(db:Session,user:schemas.UserCreateByAdmin):
     hashed_password = dep.get_password_hash(user.password)
     db_user = models.User(username=user.username, is_family=user.is_family,is_active=user.is_active,password_expired=user.password_expired,hashed_password=hashed_password)
     db.add(db_user)
@@ -71,7 +71,7 @@ def create_user_by_admin(db:Session,user:schemas.user.UserCreateByAdmin):
     db.refresh(db_user)
     return db_user
 
-def change_password(db:Session,user:schemas.user.PasswordChange):
+def change_password(db:Session,user:schemas.PasswordChange):
     db_user=db.query(models.User).filter(models.User.username==user.username).first()
     hashed_new_password = dep.get_password_hash(user.new_password)
     db_user.hashed_password=hashed_new_password
@@ -81,7 +81,7 @@ def change_password(db:Session,user:schemas.user.PasswordChange):
 
 ### 権限関係(Admin以外は要調整)
 
-def grant_admin(db:Session,user:schemas.user.User):
+def grant_admin(db:Session,user:schemas.User):
     id = int(hashids.decode(user.id)[0])
     db_admin = models.Admin(user_id=id)
     db.add(db_admin)
@@ -89,44 +89,44 @@ def grant_admin(db:Session,user:schemas.user.User):
     db.refresh(db_admin)
     return db_admin
 
-def grant_owner_of(db:Session,group:schemas.group.Group,user:schemas.user.User):
-    db_owner = models.Authority(user_id=user.id,group_id=group.id,role=schemas.authority.AuthorityRole.Owner)
+def grant_owner_of(db:Session,group:schemas.Group,user:schemas.User):
+    db_owner = models.Authority(user_id=user.id,group_id=group.id,role=schemas.AuthorityRole.Owner)
     db.add(db_owner)
     db.commit()
     db.refresh(db_owner)
     return db_owner
 
-def grant_authorizer_of(db:Session,group:schemas.group.Group,user:schemas.user.User):
-    db_authorizer = models.Authority(user_id=user.id,group_id=group.id,role=schemas.authority.AuthorityRole.Authorizer)
+def grant_authorizer_of(db:Session,group:schemas.Group,user:schemas.User):
+    db_authorizer = models.Authority(user_id=user.id,group_id=group.id,role=schemas.AuthorityRole.Authorizer)
     db.add(db_authorizer)
     db.commit()
     db.refresh(db_authorizer)
     return db_authorizer
 
-def check_admin(db:Session,user:schemas.user.User):
+def check_admin(db:Session,user:schemas.User):
     id = int(hashids.decode(user.id)[0])
     if not db.query(models.Admin).filter(models.Admin.user_id==id).first():
         return False
     return True
 
-def check_owner_of(db:Session,group:schemas.group.Group,user:schemas.user.User):
+def check_owner_of(db:Session,group:schemas.Group,user:schemas.User):
     id = int(hashids.decode(user.id)[0])
-    if not db.query(models.Authority).filter(models.Authority.user_id==id,models.Authority.group_id==group.id,models.Authority.role==schemas.authority.AuthorityRole.Owner).first():
+    if not db.query(models.Authority).filter(models.Authority.user_id==id,models.Authority.group_id==group.id,models.Authority.role==schemas.AuthorityRole.Owner).first():
         return False
     return True
 
-def check_owner(db:Session,user:schemas.user.User):
+def check_owner(db:Session,user:schemas.User):
     if not db.query(models.Authority).filter(models.Authority.user_id==user.id,models.Authority.role==schemas.AuthorityRole.Owner).first():
         return False
     return True
 
-def check_authorizer_of(db:Session,group:schemas.group.Group,user:schemas.user.User):
-    if not db.query(models.Authority).filter(models.Authority.user_id==user.id,models.Authority.group_id==group.id,models.Authority.role==schemas.authority.AuthorityRole.Authorizer).first():
+def check_authorizer_of(db:Session,group:schemas.Group,user:schemas.User):
+    if not db.query(models.Authority).filter(models.Authority.user_id==user.id,models.Authority.group_id==group.id,models.Authority.role==schemas.AuthorityRole.Authorizer).first():
         return False
     return True
 
-def check_authorizer(db:Session,user:schemas.user.User):
-    if not db.query(models.Authority).filter(models.Authority.user_id==user.id,models.Authority.role==schemas.authority.AuthorityRole.Authorizer).first():
+def check_authorizer(db:Session,user:schemas.User):
+    if not db.query(models.Authority).filter(models.Authority.user_id==user.id,models.Authority.role==schemas.AuthorityRole.Authorizer).first():
         return False
     return True
