--- app/schemas/__init__.py
@@ -1,1 +1,0 @@
-from . import authority,event,group,tag,ticket,token,user,vote
